Move custom plugins to plugin/ dir
