+++ .config/nvim/init.lua
@@ -1,83 +1,90 @@
 --=============================================================================
 -------------------------------------------------------------------------------
 --                                                                         INIT
---[[===========================================================================
-
-Ensure that Lazy.nvim (https://github.com/folke/lazy.nvim) is installed, then
-set it up. All other configs are managed as plugins through lazy.
-
------------------------------------------------------------------------------]]
+--=============================================================================
+-- Defines all the general neovim options
+-------------------------------------------------------------------------------
 
 vim.g["mapleader"] = " " -------------------------------  map <leader> to space
 
-local ensure_lazy, get_lazy_options
+---Set the default global options for the editor
+---(save/undo, indenting, autocomplete, searching, ui)
+vim.opt.errorbells = false -- disable error sounds
+vim.opt.updatetime = 50 -- shorten updatetime from 4s to 50ms
+vim.opt.timeoutlen = 300 -- shorten timeout for key combinations
+vim.opt.exrc = true -- source local .nvim.lua files
+vim.opt.splitbelow = true -- open new window below in normal split
+vim.opt.splitright = true --open new window on the right in vertical split
 
----Ensure the lazy.nvim plugin manager is installed. If not,
----intall it, then set it up and lazy load all the plugins in the
----./lua/plugins directory.
-local function load_lazy(lazy_path)
-  ensure_lazy(lazy_path)
-  vim.opt.runtimepath:prepend(lazy_path)
-  require("lazy").setup("plugins", get_lazy_options())
-end
+------------------------------------------------------------------- SAVE / UNDO
 
-function ensure_lazy(lazy_path)
-  if not vim.loop.fs_stat(lazy_path) then
-    vim.notify("Lazy.nvim not found, installing...", vim.log.levels.INFO)
-    local args = {
-      "git",
-      "clone",
-      "--filter=blob:none",
-      "--single-branch",
-      "git@github.com:folke/lazy.nvim.git",
-      lazy_path,
-    }
-    vim.notify("Running: " .. table.concat(args, " "), vim.log.levels.DEBUG)
-    vim.fn.system(args)
-    vim.notify("Lazy.nvim installed", vim.log.levels.INFO)
-  end
-end
+vim.opt.swapfile = false -- load buffers without creating swap files
+vim.opt.backup = false -- do not automatically save
+vim.opt.undofile = true -- allow undo after reoppening the file
+vim.opt.undodir = table.concat({ vim.fn.stdpath "data", "/.undo" }, "/")
 
-function get_lazy_options()
-  return {
-    defaults = {
-      lazy = true,
-      --version = "*",
-    },
-    lockfile = table.concat(
-      { vim.fn.stdpath "config", ".lazy.lock.json" },
-      "/"
-    ),
-    root = vim.fn.stdpath "data" .. "/lazy",
-    dev = {
-      dir = vim.fn.stdpath "data" .. "/lazy",
-    },
-    checker = {
-      enabled = true,
-      notify = false,
-    },
-    diff = {
-      cmd = "terminal_git",
-    },
-    performance = {
-      cache = {
-        enabled = true,
-      },
-      rtp = {
-        disabled_plugins = {
-          --"gzip",
-          "matchit",
-          "matchparen",
-          "netrwPlugin",
-          "tarPlugin",
-          "tohtml",
-          "tutor",
-          --"zipPlugin",
-        },
-      },
-    },
-    debug = false,
-  }
-end
+--------------------------------------------------------------------- INDENTING
 
-load_lazy(vim.fn.stdpath "data" .. "/lazy/lazy.nvim")
+vim.opt.tabstop = 4 -- set the width of a tab to 4
+vim.opt.softtabstop = 4 -- set the number of spaces that a tab counts for
+vim.opt.shiftwidth = 4 -- number of spaces used for each step of indent
+vim.opt.smartindent = true -- smart indent the next line
+vim.opt.expandtab = true -- use spaces in tabs
+
+------------------------------------------------------------------ AUTOCOMPLETE
+
+vim.opt.shortmess = vim.opt.shortmess + "c" -- don't give ins-completion-menu
+vim.opt.completeopt = "menu,menuone,noselect"
+
+--------------------------------------------------------------------- SEARCHING
+
+vim.opt.smartcase = true -- override ignorecase option
+vim.opt.hlsearch = false -- stop highlighting for hlsearch option
+vim.opt.incsearch = true
+vim.opt.path = ".,,**" -- search down into subfolders
+vim.opt.wildmenu = true -- display matching files with tab completion
+
+vim.opt.jumpoptions = "stack" -- make jumplist behave like stack
+
+---------------------------------------------------------------------------- UI
+
+vim.opt.background = "dark"
+vim.opt.number = true
+--vim.opt.relativenumber = true
+vim.opt.wrap = false
+vim.opt.scrolloff = 8
+vim.opt.incsearch = true
+vim.opt.title = false
+vim.opt.signcolumn = "number"
+vim.opt.showmode = false
+vim.opt.guicursor = "a:blinkon0"
+vim.opt.cursorline = true
+vim.opt.cursorcolumn = false
+--vim.opt.colorcolumn = +1
+vim.t_Co = 256
+vim.opt.termguicolors = true
+vim.cmd 'let &t_8f = "\\<Esc>[38;2;%lu;%lu;%lum"'
+vim.cmd 'let &t_8b = "\\<Esc>[48;2;%lu;%lu;%lum"'
+
+-- Set relative number and cursorline only for the active window
+vim.api.nvim_create_autocmd({ "VimEnter", "WinEnter", "BufWinEnter" }, {
+  callback = function()
+    if vim.wo.number then
+      vim.wo.relativenumber = true
+      vim.wo.cursorline = true
+    end
+  end,
+})
+vim.api.nvim_create_autocmd({ "WinLeave" }, {
+  callback = function()
+    if vim.wo.number then
+      vim.wo.relativenumber = false
+      vim.wo.cursorline = false
+    end
+  end,
+})
+
+-------------------------------------------------------------------- STATUSLINE
+-- disable statusline by default, it may then be enabled by plugins
+
+vim.opt.laststatus = 0
