--- .config/nvim/lua/plugins/options/init.lua
@@ -1,16 +1,0 @@
---=============================================================================
--------------------------------------------------------------------------------
---                                                                      OPTIONS
---=============================================================================
--- Defines all the general neovim options
--------------------------------------------------------------------------------
-
-return {
-  dir = table.concat(
-    { vim.fn.stdpath "config", "lua", "plugins", "options" },
-    "/"
-  ),
-  init = function()
-    require "options"
-  end,
-}
