+++ .config/nvim/plugin/folding.lua
@@ -5,18 +5,16 @@
 -- Fold or unfold the current context with <CR>
 --_____________________________________________________________________________
 
-local folding = {}
+local initialized = {}
 
-folding.initialized = {}
-
-function folding.fold()
+local function fold()
   local buf = vim.api.nvim_get_current_buf()
   local buftype = vim.api.nvim_buf_get_option(buf, "buftype")
   if buftype:len() > 0 then
     return
   end
-  if not folding.initialized[buf] then
-    folding.initialized[buf] = true
+  if not initialized[buf] then
+    initialized[buf] = true
     pcall(vim.api.nvim_exec, "e", true)
   end
   local ok, e = pcall(vim.api.nvim_exec, "normal! za", false)
@@ -27,11 +25,7 @@ function folding.fold()
   end
 end
 
-function folding.config()
-  vim.opt.foldmethod = "expr" -- enable folding (default 'foldmarker')
-  vim.opt.foldexpr = "nvim_treesitter#foldexpr()"
-  vim.opt.foldlevel = 9999 -- open a file fully expanded, set to
-  vim.keymap.set("n", "<CR>", folding.fold)
-end
-
-return folding
+vim.opt.foldmethod = "expr" -- enable folding (default 'foldmarker')
+vim.opt.foldexpr = "nvim_treesitter#foldexpr()"
+vim.opt.foldlevel = 9999 -- open a file fully expanded, set to
+vim.keymap.set("n", "<CR>", fold)
